apps/backend/requirements.txt
# =============================================================================
# VietSpeak AI - Backend Dependencies
# =============================================================================

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Audio Processing
librosa==0.10.1
soundfile==0.12.1
pydub==0.25.1
webrtcvad==2.0.10

# ML/AI
torch>=2.1.0
torchaudio>=2.1.0
transformers==4.37.0
openai-whisper==20231117
faster-whisper==1.0.0
numpy>=1.24.0
scipy>=1.11.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
alembic==1.13.1

# Caching & Task Queue
celery==5.3.6
redis==5.0.1

# Monitoring & Logging
structlog==24.1.0
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.26.0
aiofiles==23.2.1
